requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
bidict==0.23.1
certifi==2025.11.12
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
docutils==0.22.4
fastapi==0.127.1
frozenlist==1.8.0
h11==0.16.0
html-sanitizer==2.6.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
nicegui==3.4.1
orjson==3.11.5
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.21
python-socketio==5.16.0
PyYAML==6.0.3
rich==14.2.0
simple-websocket==1.1.0
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
